2nd Bom Animation
Re-worked Bom Animation and named the instance Bom 2. Bom is now Kinematic and put the loop on the anim to compare. The first bom animation the piece were just flying away while rotating and shrinking. Bom2 animation was made as the piece of the Bom actually split in smaller pieces.